backend/requirements.txt
alembic==1.16.5
bcrypt==4.3.0
blinker==1.9.0
Brotli==1.1.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
cssselect2==0.8.0
Flask==3.1.2
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fonttools==4.59.2
greenlet==3.2.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
mysql-connector-python==9.4.0
pillow==11.3.0
pycparser==2.22
pydyf==0.11.0
PyJWT==2.10.1
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
qrcode==8.2
requests==2.32.5
six==1.17.0
SQLAlchemy==2.0.43
tinycss2==1.4.0
tinyhtml5==2.0.0
tomli==2.2.1
typing_extensions==4.15.0
urllib3==2.5.0
weasyprint==66.0
webencodings==0.5.1
Werkzeug==3.1.3
WTForms==3.2.1
zopfli==0.2.3.post1
